Fix ‘This product doesn't exist.’
Currently, when you import the attributes the backend_type is not updated/inserted into the eav_attribute table. Because of that the default backend_type ‘static’ is used. But with backend_type ‘static’ Magento tries to get the information from catalog_product_entity table where none of these attributes exist as column, so an SQL error occures.
By adding backend_type and some more, which are explicitly mapped in matchType(), Magento searches in the right tabel i.e. catalog_product_entity_varchar for the attribute value.